+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Check that the circuits submitted to the offline simulators can be converted to the AQT API (#68)
+* Update the user guide and improve the API reference consistency (#72)
+* Add quickstart examples for the Qiskit.org homepage (#73)
+
 ## qiskit-aqt-provider v0.17.0
 
 * Merge community and AQT versions (#61)